@@ -17,19 +17,20 @@
 package net.tridentsdk.server.world;
 
 import com.google.common.collect.Lists;
-import io.netty.util.internal.chmv8.ConcurrentHashMapV8;
 import net.tridentsdk.concurrent.HeldValueLatch;
 import net.tridentsdk.docs.AccessNoDoc;
 import net.tridentsdk.util.TridentLogger;
+import net.tridentsdk.world.Chunk;
 import net.tridentsdk.world.ChunkLocation;
 
 import java.util.Collection;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 @AccessNoDoc
 class ChunkCache {
-    private final ConcurrentMap<ChunkLocation, HeldValueLatch<TridentChunk>> cachedChunks = new ConcurrentHashMapV8<>();
+    private final ConcurrentMap<ChunkLocation, HeldValueLatch<TridentChunk>> cachedChunks = new ConcurrentHashMap<>();
     private final TridentWorld world;
 
     public ChunkCache(TridentWorld world) {
@@ -37,21 +38,20 @@ class ChunkCache {
     }
 
     public void put(ChunkLocation location, TridentChunk chunk) {
-        HeldValueLatch<TridentChunk> latch = cachedChunks.get(location);
-        if (latch == null)
-            latch = HeldValueLatch.create();
-        if (!latch.hasValue())
-            latch.countDown(chunk);
-        else {
-            latch = HeldValueLatch.create();
-            latch.countDown(chunk);
+        HeldValueLatch<TridentChunk> value = cachedChunks.get(location);
+        if (value == null) {
+            HeldValueLatch<TridentChunk> latch = HeldValueLatch.create();
+            value = cachedChunks.putIfAbsent(location, latch);
+            if (value == null) {
+                value = latch;
+                value.countDown(chunk);
+            }
         }
-
-        cachedChunks.put(location, latch);
     }
 
     public TridentChunk get(ChunkLocation location, boolean gen) {
-        while (true) {
+        Chunk chunk = null;
+        while (chunk == null) {
             HeldValueLatch<TridentChunk> value = cachedChunks.get(location);
             if (value == null) {
                 if (!gen) return null;
@@ -68,11 +68,13 @@ class ChunkCache {
                 return null;
 
             try {
-                return value.await();
+                chunk = value.await();
             } catch (InterruptedException e) {
                 e.printStackTrace();
             }
         }
+
+        return (TridentChunk) chunk;
     }
 
     public void retain(Set<ChunkLocation> set) {
@@ -81,12 +83,12 @@ class ChunkCache {
             return;
         }
 
-        for (ChunkLocation l : keys()) {
-            if (!set.contains(l)) {
-                HeldValueLatch<TridentChunk> latch = cachedChunks.remove(l);
-                latch.get().clear();
-            }
-        }
+        keys().stream()
+                .filter(set::contains)
+                .forEach((l) -> {
+                    HeldValueLatch<TridentChunk> latch = cachedChunks.remove(l);
+                    latch.get().clear();
+                });
     }
 
     public Set<ChunkLocation> keys() {
@@ -95,15 +97,16 @@ class ChunkCache {
 
     public Collection<TridentChunk> values() {
         Collection<TridentChunk> chunks = Lists.newArrayList();
-        for (HeldValueLatch<TridentChunk> chunk : cachedChunks.values()) {
-            if (chunk.hasValue()) {
-                try {
-                    chunks.add(chunk.await());
-                } catch (InterruptedException e) {
-                    TridentLogger.error(e);
-                }
-            }
-        }
+
+        cachedChunks.values().stream()
+                .filter(HeldValueLatch::hasValue)
+                .forEach((c) -> {
+                    try {
+                        chunks.add(c.await());
+                    } catch (InterruptedException e) {
+                        TridentLogger.error(e);
+                    }
+                });
 
         return chunks;
     }
